Add Unicode 8.0 codepoints
The provided awk script was used to extract codepoints and names from
UnicodeData.txt.  For control characters that don't have a primary name,
the alias name is taken instead.

The result was manually compared to the prior list.  Most lines are
simple additions.  For cases where the names came out different,
preference was given to the new Unicode data.  In a few cases, the old
list had names that the new list didn't, and those are still retained.
(e.g. U+0080 = 'PADDING CHARACTER', and all the '<reserved>' entries.)